Add low throughput mode to BackPressureHandler
Enable partial acquires

Until now the solution was optimized for high throughput scenarios, which would result in many unnecessary polls being made for idle moments.

Now when no messages are returned from a poll and there are no inflight messages the application will fall back to a low throughput mode where only one poll is made at a time. When a message is returned, it goes back to the high throughput mode. This automatic switching can be disabled via ContainerOptions.

Also, now if a whole batch of permits cannot be acquired, the application will try to acquire a smaller batch, so it doesn't have to wait for an entire batch to complete before polling for more messages.